Move non wallet buttons from left sidebar
